ibmi: Remove PASE differences for SQLBindParameter when data-at-exec
The IBM i CLI docs say that when using SQL_DATA_AT_EXEC, ParameterValuePtr
should be a pointer to a 32-bit value. This 32-bit value is returned on
the SQLParamData call. Either this has been modified when running under
PASE or this was wrong to begin with. SQLParamData returns the pointer
that it was given on SQLBindParameter. Might need to double check
for 64-bit, but there's no 64-bit python support in PASE yet.