ibmi: Convert column size for binary types to match DB2 Connect CLI
IBM i CLI returns the binary size instead of the string size for
the column size. eg. SQL_INTEGER will give 4 bytes instead of 10
characters, since an SQL_INTEGER is a 4 byte (32-bit) integer.

We could just leave this as is and adjust the tests, but when binding
columns, _python_ibm_db_bind_column_helper uses this value to determine
how many bytes to allocate for the output. This causes truncation
when run on IBM i.