ibmi: Fix callproc handling due to bad SQL_NO_TOTAL value
callproc checks whether the indicator contains SQL_NULL_DATA or
SQL_NO_TOTAL, but SQL_NO_TOTAL is defined as 0 on IBM i (which is
a valid length). This erroneous mistake in the header file causes
output parameters for binary (SQLINTEGER, etc...) types to fail,
since the length is not needed and is set to 0.

I have updated the header file and it will be included in a future
release of IBM i to define SQL_NO_TOTAL as -4, following DB2 Connect
and ODBC.