ibmi: Implement ibm_db.get_num_result properly on IBM i
IBM i CLI does not implement SQL_DIAG_CURSOR_ROW_COUNT and previously
I had used SQLRowCount instead, but this is not the same. Instead,
I've implemented SQL_DIAG_CURSOR_ROW_COUNT in IBM i CLI, which will
be enabled in a future PTF and the include file will be updated in
a future release.