ibmi: Implement execute_many better on IBM i
IBM i CLI does not support chaining, but since we're local to the
box, we don't really need to worry about network performance that
chaining is supposed to improve. Instead, we just execute all the
statements individually.

Since they are not being executed together, the row count will
always be 1, though. Instead, we cache the row count based on the
number of times we executed the statement minus any statements that
were in error. We then return this cached value. We need to cache
it so that if an application calls ibm_db.num_rows it will return
the same value. On other platforms, we'll continue to use SQLRowCount.

In the future, we could think about using block insert to mimic
chaining or implement chaining in the IBM i CLI interface. For now,
it does not seem worth the benefit.