ibmi: Mimic DB2 Connect field width values on IBM i
On IBM i CLI, SQL_DESC_LENGTH returns the number of bytes used to
store the data. So SQLINTEGER would return 4, for example. DB2 Connect,
at least when the version is set to ODBC version 3, returns the character
length of the field. I may push this down in to CLI at some point.