ibm_db.statistics should return NULL instead of Py_False when an exception is set
Unpredictable behavior will occur if the exception is not cleared or NULL is not
returned, since the exception object will be checked at some later time when it
is not expected.

If we really should return Py_False instead, the exception should be cleared.
However, that will make it more difficult to determine what the problem is, since
there is no statement object to pass to ibm_db.stmt_errormsg.