ibmi: Work around IBM i CLI returning CHAR for BINARY types
IBM i DB2 supports (VAR)CHAR FOR BIT DATA and (VAR)CHAR CCSID(65535)
which are essentially (VAR)BINARY columns, but returns SQL_CHAR as
the type from SQLDescribeCol and SQLColAttribute. When we bind as
SQL_C_CHAR, CLI tries to convert the data from binary to UTF-8,
which is impossible. Instead, we should treat these as binary types.
This also affects ROWID columns as well.

At some point, this will be fixed in CLI and this code can be removed.