[ibmi] Make getUnicodeDataAsSQLTCHAR work on Python 2
Python 2 doesn't have PyUnicode_AsUTF8AndSize, so we have to three-
step it using PyUnicode_AsUTF8String, PyUnicode_AsUTF8String, and
PyString_AsString.

This also fixes a bug in the Python 3 version where we return
utf8 temp string instead of the copy we made, resulting in a memory
leak since we free the wrong buffer.