inc/dec bound Python objects and ensure cache is cleared properly
When binding a Python object to a parameter, we store a pointer to
the Python object, but we do not increment the refcount. If the
object loses all references, it will be automatically cleaned up
by the interpreter and the results of our execute are undeterminable
and can result in Segmentation Faults.

Additionally, while adding this code I found that code to clear the
param cache was in various places and was inconsistent, resulting
in possible memory leaks due to unfreed memory. This has been
consolidated to a new function: _python_ibm_db_clear_param_cache

Fixes #216