[ibmi] Fix timestamp handling
The IBM i CLI interface is unable to map TIMESTAMP(n) to TIMESTAMP(m)
where n != m. Thus when we must specify the described length exactly.
The resulting TIMESTAMP comes back as a string without a NULL terminator,
but the scale and length are known.

DB2 on IBM i (at 7.2 or above) supports up to 12 seconds of precision, but
the TIMESTAMP_STRUCT supports up to 9 digits of precision. For now, we just
ignore the possible 3 trailing digits. If TIMESTAMP has fewer than 9 digits
of precision, we adjust to 9 digits. All this is really for nought, though,
since the Python datetime module only supports up to 6 digits of precision.
See this bug for more details: https://bugs.python.org/issue15443