[ibmi] Allow passing zero parameters to ibm_db_dbi.(p)connect on IBM i
On IBM i you can connect through SQLConnect(*LOCAL, NULL, NULL) to
connect as the local user to the local database. There used to be
support for passing in empty strings, similarly, but that support was
removed. Since passing in empty strings is invalid, we treat those
the same as None. This also has the benefit of being the default
parameters for connect/pconnect, so if the user specifies no
parameteters, they will connect just fine.

Also add the missing default parameter for dsn on connect(), which
completes the support.