removed broken check when adding M2M field
The implementation brought up an Exception when adding a new M2M-field with selfwritten through-Model. That happened because auto_create is False in this case, it is a notnull-field so it was trying to alter the column to not null throwing "Column, attribute, or period [M2M-field-name] is not defined in [table-name]".
All the logic that determines if it is an M2M that has to be auto/manually created is already done in the super().add_field call. A M2M-field never adds a new column but always a new table. It does not matter if it's auto/manually created. So all the notnull, unique and p_key stuff is never relevant.