Fix bug in setup.py where wrong cli file could be downloaded (#325)
* Fix the broken links

I read and certify the DCO; as an employee of IBM, they own my work anyway.

* Fix bug in setup.py where wrong cli file could be downloaded

I work for IBM; this work is theirs, so I accept the contributor guidelines.

Anyway, there’s a bug in the setup for ibm_db.  Specifically, when you’re trying to install the command line library.  You have function that looks through everything that’s returned from os.uname() to determine which one to download.  That function looks through all the items returned, which includes (sysname,nodename, release, version, machine).  The problem is that when you’re checking to determine which libraries to download for linux, you look at all those fields, and right now, the release string for RHEL zLinux is '3.10.0-862.el7.s390x’, so if you try installing ibm_db on zLinux, it downloads the cli_libraries for the x86 platform and the install fails.  What I’m proposing you do instead is only look at the machine field to determine the architecture.  I don’t have access to all those types of machines to validate that this is the correct solution, but it does work for zLinux.  I’m assuming you have a place you can validate the other platforms.

* Changing check for the SPARC architecture

Signed-off-by: Kevin McKenzie <kmckenzi@us.ibm.com>